refactor: 폴더 구조 개선 및 파일 이동
- components 폴더 생성으로 UI 관련 컴포넌트 파일 이동
