+++ src/components/TodoList.jsx
@@ -1,8 +1,7 @@
 import React from "react";
-import TodoItem from "./TodoItem";
+import TodoItem from "../TodoItem";
 
 // 할 일 목록을 표시하는 컴포넌트
-
 const TodoList = ({ todos, toggleComplete, deleteTodo }) => {
   return (
     <ul>
